requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
deprecation==2.1.0
Django==5.0.7
djangorestframework==3.15.2
ecdsa==0.19.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jwcrypto==1.5.6
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
python-jose==3.3.0
python-keycloak==4.2.3
python-keycloak-client==0.2.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
